vec4i_t for viewport.

+++ src/client/renderer/r_main.c
@@ -221,7 +221,7 @@ static void R_UpdateUniforms(const r_view_t *view) {
 	if (view) {
 		r_uniforms.block.viewport = view->viewport;
 
-		const float aspect = view->viewport.z / view->viewport.w;
+		const float aspect = view->viewport.z / (float) view->viewport.w;
 
 		const float ymax = tanf(Radians(view->fov.y));
 		const float ymin = -ymax;
